fix(core): take restricted.decorator off every import cycle
restricted.decorator sat on TWO runtime cycles at once:

  restricted.decorator -> db.helper -> input.helper -> restricted.decorator
  restricted.decorator -> tenant/core-tenant.helpers -> config.service
                       -> input.helper -> restricted.decorator

input.helper therefore evaluated while restricted.decorator was still
initializing. Nothing crashed, but only because every cross-cycle dereference
happened to sit inside a function body. A single top-level line in input.helper
— a module-level alias of checkRestricted, an @Restricted-decorated class,
design:type metadata — would have thrown under SWC -> CommonJS:

  ReferenceError: Cannot access 'checkRestricted' before initialization

…in the file that drives field-level access control, on a compiler the default
build never runs, with a fully green test suite.

Two leaf extractions remove both edges:

  - id.helper.ts    <- the ID cluster (equalIds, getIncludedIds, getStringIds,
                       getObjectIds) out of db.helper. Depends only on mongoose
                       Types, lodash and a type.
  - clone.helper.ts <- clone / deepFreeze out of input.helper, so config.service
                       can reach them without importing input.helper.

Both are true leaves (no framework imports), and db.helper / input.helper
re-export their former members, so every existing import path keeps working.

Removing one edge is not the same as removing the cycle: the id.helper
extraction alone left the config.service path fully intact. restricted.decorator
now appears in ZERO cycles; repo-wide cycles drop from 10 to 6, and most of the
remainder are type-only imports that madge reports but both compilers erase.

Its three exports are additionally converted from `const` arrows to hoisted
function declarations. With the cycles gone this is defense in depth: hoisted
declarations are initialized before any module body runs, so they stay TDZ-immune
whatever the import graph does next.
